Renamed commsmanager
